+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Language picker in auth page
 - Custom Drinks can now be deleted
+- BeerWithMe data import
 
 ### Changed
 - Branding